CameraX functionality fully implemented
CameraX can view through the rear-facing camera, resolution is low but can be changed if necessary. CameraX can take a photo and returns a ProxyImage image which is then to be passed to the barcode class. We may need to convert the ImageProxy image object to another if the barcode class requires us to use a different format/class.